# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FREGE AI – FregeOS Agricultural Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /farmgpt/conversations:
    get:
      operationId: getFarmGptConversations
      tags: [farmgpt]
      summary: List FarmGPT conversation history
      responses:
        "200":
          description: Conversations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createFarmGptConversation
      tags: [farmgpt]
      summary: Start a new FarmGPT conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /farmgpt/conversations/{conversationId}/messages:
    get:
      operationId: getFarmGptMessages
      tags: [farmgpt]
      summary: Get messages in a conversation
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendFarmGptMessage
      tags: [farmgpt]
      summary: Send a message to FarmGPT
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent with AI reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePair"

  # ─── FARM MANAGEMENT ───────────────────────────────────────────────────────
  /farms:
    get:
      operationId: getFarms
      tags: [farm]
      summary: List all farms for the farmer
      responses:
        "200":
          description: Farms list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Farm"
    post:
      operationId: createFarm
      tags: [farm]
      summary: Create a new farm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmInput"
      responses:
        "201":
          description: Created farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
  /farms/{farmId}:
    get:
      operationId: getFarm
      tags: [farm]
      summary: Get a specific farm
      parameters:
        - in: path
          name: farmId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Farm details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
    patch:
      operationId: updateFarm
      tags: [farm]
      summary: Update farm details
      parameters:
        - in: path
          name: farmId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FarmUpdate"
      responses:
        "200":
          description: Updated farm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Farm"
...
  /market/listings:
    get:
      operationId: getMarketListings
      tags: [market]
      summary: List all market listings
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [crop, livestock, input, equipment]
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketListing"
    post:
      operationId: createMarketListing
      tags: [market]
      summary: Create a new market listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketListingInput"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketListing"
  /market/listings/{listingId}:
    get:
      operationId: getMarketListing
      tags: [market]
      summary: Get listing details
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketListing"
    delete:
      operationId: deleteMarketListing
      tags: [market]
      summary: Delete a listing
      parameters:
        - in: path
          name: listingId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Conversation:
      type: object
      required: [id, title, language, messageCount, lastMessageAt, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        language:
          type: string
          enum: [english, hausa, yoruba, igbo, fulfulde]
        messageCount:
          type: integer
        lastMessageAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ConversationInput:
      type: object
      required: [title, language]
      properties:
        title:
          type: string
        language:
          type: string
          enum: [english, hausa, yoruba, igbo, fulfulde]
    Message:
      type: object
      required: [id, conversationId, role, content, createdAt]
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
    MessagePair:
      type: object
      required: [userMessage, assistantMessage]
      properties:
        userMessage:
          $ref: "#/components/schemas/Message"
        assistantMessage:
          $ref: "#/components/schemas/Message"

    # ─── FARM ─────────────────────────────────────────────────────────────────
    Farm:
      type: object
      required: [id, name, sizeHectares, state, lga, farmType, healthScore, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sizeHectares:
          type: number
        state:
          type: string
        lga:
          type: string
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        farmType:
          type: string
          enum: [crop, livestock, mixed, aquaculture]
        healthScore:
          type: number
        soilType:
          type: ["string", "null"]
        irrigationType:
          type: ["string", "null"]
        createdAt:
          type: string
    FarmInput:
      type: object
      required: [name, sizeHectares, state, lga, farmType]
      properties:
        name:
          type: string
        sizeHectares:
          type: number
        state:
          type: string
        lga:
          type: string
        lat:
          type: number
        lng:
          type: number
        farmType:
          type: string
          enum: [crop, livestock, mixed, aquaculture]
        soilType:
          type: string
        irrigationType:
          type: string
    FarmUpdate:
      type: object
      properties:
        name:
          type: string
        sizeHectares:
          type: number
        soilType:
          type: string
        irrigationType:
          type: string
...
    MarketListing:
      type: object
      required: [id, farmerId, farmerName, type, title, description, priceNgn, quantity, unit, state, lga, status, createdAt]
      properties:
        id:
          type: integer
        farmerId:
          type: integer
        farmerName:
          type: string
        farmerAvatar:
          type: ["string", "null"]
        farmerVerified:
          type: boolean
        type:
          type: string
          enum: [crop, livestock, input, equipment]
        title:
          type: string
        description:
          type: string
        priceNgn:
          type: number
        quantity:
          type: number
        unit:
          type: string
        state:
          type: string
        lga:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, sold, expired]
        createdAt:
          type: string
    MarketListingInput:
      type: object
      required: [type, title, description, priceNgn, quantity, unit]
      properties:
        type:
          type: string
          enum: [crop, livestock, input, equipment]
        title:
          type: string
        description:
          type: string
        priceNgn:
          type: number
        quantity:
          type: number
        unit:
          type: string
        imageUrl:
          type: string
...