# lib/api-zod/src/generated/types/marketListing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FREGE AI – FregeOS Agricultural Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketListingStatus } from './marketListingStatus';
import type { MarketListingType } from './marketListingType';

export interface MarketListing {
  id: number;
  farmerId: number;
  farmerName: string;
  /** @nullable */
  farmerAvatar?: string | null;
  farmerVerified?: boolean;
  type: MarketListingType;
  title: string;
  description: string;
  priceNgn: number;
  quantity: number;
  unit: string;
  state: string;
  lga: string;
  /** @nullable */
  imageUrl?: string | null;
  status: MarketListingStatus;
  createdAt: string;
}